Binance_Futures_Tradings_Bot/requirements.txt
requests>=2.28.0
pandas>=1.5.0
numpy>=1.24.0
ta>=0.10.0
pytest>=7.1.0
pyyaml>=6.0